+++ backend/yigil-api/src/main/java/kr/co/yigil/favor/domain/Favor.java
@@ -25,7 +25,7 @@ public class Favor {
     private Member member;
 
     @ManyToOne
-    @JoinColumn(name = "post_id")
+    @JoinColumn(name = "travel_id")
     private Travel travel;
 
     public Favor(final Member member, final Travel travel) {
